Parser: split file
