+++ projects/06/Assembler/global.h
@@ -1,0 +1,12 @@
+/**
+    Global Constants and Header
+*/
+
+#ifndef __GLOBAL_H__
+#define __GLOBAL_H__
+
+#include <iostream>
+#include <fstream>
+#include <string>
+
+#endif
